Moved the backend __init__ into another file. (#12086)
* Moved the backend __init__ into another file.

* Used the private variable of load_backend.
