+++ tests/keras/utils/data_utils_test.py
@@ -25,6 +25,7 @@ from keras.utils.data_utils import _hash_file
 from keras.utils.data_utils import get_file
 from keras.utils.data_utils import validate_file
 from keras import backend as K
+from keras.backend import load_backend
 
 pytestmark = pytest.mark.skipif(
     six.PY2 and 'TRAVIS_PYTHON_VERSION' in os.environ,
@@ -88,21 +89,21 @@ def test_data_utils(in_tmpdir):
     path = get_file(dirname, origin, untar=True)
     filepath = path + '.tar.gz'
     data_keras_home = os.path.dirname(os.path.dirname(os.path.abspath(filepath)))
-    assert data_keras_home == os.path.dirname(K._config_path)
+    assert data_keras_home == os.path.dirname(load_backend._config_path)
     os.remove(filepath)
 
     _keras_home = os.path.join(os.path.abspath('.'), '.keras')
     if not os.path.exists(_keras_home):
         os.makedirs(_keras_home)
     os.environ['KERAS_HOME'] = _keras_home
-    reload_module(K)
+    reload_module(load_backend)
     path = get_file(dirname, origin, untar=True)
     filepath = path + '.tar.gz'
     data_keras_home = os.path.dirname(os.path.dirname(os.path.abspath(filepath)))
-    assert data_keras_home == os.path.dirname(K._config_path)
+    assert data_keras_home == os.path.dirname(load_backend._config_path)
     os.environ.pop('KERAS_HOME')
     shutil.rmtree(_keras_home)
-    reload_module(K)
+    reload_module(load_backend)
 
     path = get_file(dirname, origin, untar=True)
     filepath = path + '.tar.gz'
